Refactor app, add wolf howl sound, and configure Windows EXE build
- Refactored `Main.java` into `BackgroundSnap`, `CameraHandler`, and `ImageProcessor`.
- Added `wolf-howl.mp3` and implemented playback using JLayer.
- Configured `launch4j` gradle plugin for Windows EXE generation.
- Added unit tests for image processing and resource loading.
- Ensured Swing UI updates run on the EDT.
